+++ components/library/src/library.rs
@@ -81,7 +81,7 @@ impl Library {
     /// Find out the direct subsections of each subsection if there are some
     /// as well as the pages for each section
     pub fn populate_sections(&mut self) {
-        let mut grandparent_paths: HashMap<PathBuf, Vec<PathBuf>> = HashMap::new();
+        let mut grandparent_paths: HashMap<PathBuf, Vec<_>> = HashMap::new();
 
         for section in self.sections.values_mut() {
             if let Some(ref grand_parent) = section.file.grand_parent {
@@ -99,6 +99,7 @@ impl Library {
             let parent_section_path = page.file.parent.join("_index.md");
             if let Some(section_key) = self.paths_to_sections.get(&parent_section_path) {
                 self.sections.get_mut(*section_key).unwrap().pages.push(key);
+                page.parent_section = Some(*section_key);
             }
         }
 
@@ -109,14 +110,22 @@ impl Library {
         for (key, section) in &self.sections {
             sections_weight.insert(key, section.meta.weight);
         }
-        for section in self.sections.values_mut() {
-            if let Some(paths) = grandparent_paths.get(&section.file.parent) {
-                section.subsections = paths
-                    .iter()
-                    .map(|p| sections[p])
-                    .collect::<Vec<_>>();
-                section.subsections
-                    .sort_by(|a, b| sections_weight[a].cmp(&sections_weight[b]));
+
+        for (grandparent, children) in &grandparent_paths {
+            let mut subsections = vec![];
+            let grandparent_path = grandparent.join("_index.md");
+
+            if let Some(ref mut section) = self.get_section_mut(&grandparent_path) {
+                subsections = children.iter().map(|p| sections[p]).collect();
+                subsections.sort_by(|a, b| sections_weight[a].cmp(&sections_weight[b]));
+                section.subsections = subsections.clone();
+            }
+
+            // Only there for subsections so we must have a parent section
+            for key in &subsections {
+                if let Some(ref mut subsection) = self.sections.get_mut(*key) {
+                    subsection.parent_section = Some(sections[&grandparent_path]);
+                }
             }
         }
     }
@@ -219,6 +228,11 @@ impl Library {
         None
     }
 
+    /// Only used in tests
+    pub fn get_section_key(&self, path: &PathBuf) -> Option<&Key> {
+        self.paths_to_sections.get(path)
+    }
+
     pub fn get_section(&self, path: &PathBuf) -> Option<&Section> {
         self.sections.get(self.paths_to_sections.get(path).cloned().unwrap_or_default())
     }
@@ -231,6 +245,14 @@ impl Library {
         self.sections.get(key).unwrap()
     }
 
+    pub fn get_section_mut_by_key(&mut self, key: Key) -> &mut Section {
+        self.sections.get_mut(key).unwrap()
+    }
+
+    pub fn get_section_path_by_key(&self, key: Key) -> &str {
+        &self.get_section_by_key(key).file.relative
+    }
+
     pub fn get_page(&self, path: &PathBuf) -> Option<&Page> {
         self.pages.get(self.paths_to_pages.get(path).cloned().unwrap_or_default())
     }
@@ -241,7 +263,6 @@ impl Library {
 
     pub fn remove_section(&mut self, path: &PathBuf) -> Option<Section> {
         if let Some(k) = self.paths_to_sections.remove(path) {
-            // TODO: delete section from parent subsection if there is one
             self.sections.remove(k)
         } else {
             None
@@ -250,7 +271,6 @@ impl Library {
 
     pub fn remove_page(&mut self, path: &PathBuf) -> Option<Page> {
         if let Some(k) = self.paths_to_pages.remove(path) {
-            // TODO: delete page from all parent sections
             self.pages.remove(k)
         } else {
             None